test: added integration tests for EventController
